fixed more compiler warnings
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@4951 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ common/include/pcl/common/impl/centroid.hpp
@@ -847,7 +847,7 @@ pcl::demeanPointCloud (const pcl::PointCloud<PointT> &cloud_in,
   }
   else
   {
-    cloud_out.width    = indices.size ();
+    cloud_out.width    = static_cast<uint32_t> (indices.size ());
     cloud_out.height   = 1;
   }
   cloud_out.points.resize (indices.size ());
